requirements.txt
python-jobspy
pandas
scikit-learn
fastapi
uvicorn
jinja2
python-multipart
requests
markdown
pytest
openpyxl
